Ajustes em reservas.

+++ Comandas.Api/Comandas.Api/Controllers/ReservasController.cs
@@ -51,8 +51,25 @@ namespace Comandas.Api.Controllers
             {
                 return BadRequest();
             }
-
             _context.Entry(reserva).State = EntityState.Modified;
+
+            // -----------
+            var novaMesa = await _context.Mesas
+                    .FirstOrDefaultAsync(m => m.NumeroMesa == reserva.NumeroMesa);
+            if (novaMesa is null)
+                return BadRequest("Mesa não encontrada.");
+            novaMesa.SituacaoMesa = (int)SituacaoMesa.Reservada;
+
+            // consulta dados da reserva original
+            var reservaOriginal = await _context.Reservas.AsNoTracking()
+                .FirstOrDefaultAsync(r => r.Id == id);
+            // consulta numero da mesa original
+            var numeroMesaOriginal = reservaOriginal!.NumeroMesa;
+            // consulta a mesa original
+            var mesaOriginal = await _context.Mesas
+                    .FirstOrDefaultAsync(m => m.NumeroMesa == numeroMesaOriginal);
+            mesaOriginal!.SituacaoMesa = (int)SituacaoMesa.Livre;
+            // -----------
 
             try
             {
@@ -79,6 +96,25 @@ namespace Comandas.Api.Controllers
         public async Task<ActionResult<Reserva>> PostReserva(Reserva reserva)
         {
             _context.Reservas.Add(reserva);
+
+            //atualizar o status da mesa para reservada
+            //Consultando a mesa pelo numero
+            var mesa = await _context.Mesas
+                .FirstOrDefaultAsync(m => m.NumeroMesa == reserva.NumeroMesa);
+            // se mesa encontrada
+            if (mesa is null)
+            {
+                return BadRequest("Mesa não encontrada.");
+            }
+            //se a mesa não encontrada
+            if(mesa is not null)
+            {
+                if(mesa.SituacaoMesa != (int)SituacaoMesa.Livre)
+                {
+                    return BadRequest("Mesa não está disponível para reserva.");
+                }
+                mesa.SituacaoMesa = (int)SituacaoMesa.Reservada;                
+            }
             await _context.SaveChangesAsync();
 
             return CreatedAtAction("GetReserva", new { id = reserva.Id }, reserva);
@@ -91,9 +127,19 @@ namespace Comandas.Api.Controllers
             var reserva = await _context.Reservas.FindAsync(id);
             if (reserva == null)
             {
-                return NotFound();
+                return NotFound("Reserva não encontrada");
             }
-
+            // ---------------------
+            //consultar a mesa
+            var mesa = await _context.Mesas
+                .FirstOrDefaultAsync(m => m.NumeroMesa == reserva.NumeroMesa);
+            if (mesa is null)
+            {
+                return BadRequest("Mesa não encontrada.");
+            }
+            //atualizar o status da mesa para livre
+            mesa.SituacaoMesa = (int)SituacaoMesa.Livre; // (int) converte o enum para int
+            //----------------------
             _context.Reservas.Remove(reserva);
             await _context.SaveChangesAsync();
